Use explicit initializer in participant declarations service
